+++ src/interrogate/cli.py
@@ -135,7 +135,20 @@ from interrogate import utils
     is_flag=True,
     default=False,
     show_default=True,
-    help="Ignore methods with property setter/getter decorators.",
+    help="Ignore methods with property setter/getter/deleter decorators.",
+)
+@click.option(
+    "-r",
+    "--ignore-regex",
+    type=str,
+    default=(),
+    multiple=True,
+    metavar="STR",
+    callback=utils.parse_regex,
+    help=(
+        "Regex identifying class, method, and function names to ignore. "
+        "Multiple `-r/--ignore-regex` invocations supported."
+    ),
 )
 @click.option(
     "-S",
@@ -157,17 +170,11 @@ from interrogate import utils
     ),
 )
 @click.option(
-    "-r",
-    "--ignore-regex",
-    type=str,
-    default=(),
-    multiple=True,
-    metavar="STR",
-    callback=utils.parse_regex,
-    help=(
-        "Regex identifying class, method, and function names to ignore. "
-        "Multiple `-r/--ignore-regex` invocations supported."
-    ),
+    "--pyi",
+    is_flag=True,
+    default=False,
+    show_default=True,
+    help="Include stub files that end in `.pyi`.",
 )
 @click.option(
     "-w",
@@ -180,6 +187,18 @@ from interrogate import utils
     help=(
         "Regex identifying class, method, and function names to include. "
         "Multiple `-w/--whitelist-regex` invocations supported."
+    ),
+)
+@click.option(
+    "--style",
+    type=click.Choice(["sphinx", "google"]),
+    default="sphinx",
+    show_default=True,
+    help=(
+        "Style of docstrings to honor.\nUsing `google` will consider a class "
+        "and its `__init__` method both covered if there is either a class-"
+        "level docstring, or an `__init__` method docstring, instead of "
+        "enforcing both. Mutually exclusive with `-i`/`--ignore-init` flag."
     ),
 )
 @click.option(
@@ -301,9 +320,13 @@ def main(ctx, paths, **kwargs):
     .. versionadded:: 1.4.0 ``--ignore-setters``
     .. versionadded:: 1.5.0 ``--omit-covered-files``
     .. versionadded:: 1.5.0 ``--badge-style``
+    .. versionadded:: 23.0 ``--pyi``
+    .. versionadded:: 23.0 ``--style=sphinx|google``
 
     .. versionchanged:: 1.3.1 only generate badge if results change from
         an existing badge.
+    .. versionchanged:: 23.0 include property deleters when ignoring all
+        property decorators (--ignore-property-decorators)
     """
     gen_badge = kwargs["generate_badge"]
     if kwargs["badge_format"] is not None and gen_badge is None:
@@ -315,6 +338,14 @@ def main(ctx, paths, **kwargs):
         raise click.BadParameter(
             "The `--badge-style` option must be used along with the `-g/"
             "--generate-badge option."
+        )
+    if kwargs["style"] == "google" and kwargs["ignore_init_method"]:
+        raise click.BadOptionUsage(
+            option_name="style",
+            message=(
+                "The `--ignore-init-method` flag is mutually exclusive with "
+                "`--style google`."
+            ),
         )
     if not paths:
         paths = (os.path.abspath(os.getcwd()),)
@@ -342,7 +373,9 @@ def main(ctx, paths, **kwargs):
         ignore_property_setters=kwargs["ignore_setters"],
         ignore_property_decorators=kwargs["ignore_property_decorators"],
         include_regex=kwargs["whitelist_regex"],
+        include_stubs=kwargs["pyi"],
         omit_covered_files=kwargs["omit_covered_files"],
+        docstring_style=kwargs["style"],
     )
     interrogate_coverage = coverage.InterrogateCoverage(
         paths=paths,